Move all the SSL stuff into a separate Component, using a variant to select the implementation.
