--- Sming/Components/AxtlsSsl/README.rst
@@ -1,4 +1,0 @@
-AxTLS SSL
-==========
-
-Provides implementation for Secure Socket Layer (SSL) based on axTLS.
