--- Sming/Components/AxtlsSsl/component.mk
@@ -1,2 +1,0 @@
-# SSL implementation based on axTLS
-COMPONENT_DEPENDS := axtls-8266
