--- Sming/Components/AxtlsSsl/include/AxtlsSsl.h
@@ -1,19 +1,0 @@
-/****
- * Sming Framework Project - Open Source framework for high efficiency native ESP8266 development.
- * Created 2015 by Skurydin Alexey
- * http://github.com/SmingHub/Sming
- * All files of the Sming Core are provided under the LGPL v3 license.
- *
- * AxtlsSsl.h
- *
- * @author: 2019 - Slavey Karadzhov <slav@attachix.com>
- *
- ****/
-
-#include <Network/Ssl/Factory.h>
-
-class AxtlsFactory : public Ssl::Factory
-{
-public:
-	Ssl::Context* createContext() override;
-};
