+++ Sming/Components/axtls-8266/component.mk
@@ -1,10 +1,10 @@
-COMPONENT_VARS			:= SSL_DEBUG
-SSL_DEBUG				?= 0
+CACHE_VARS			:= SSL_DEBUG
+SSL_DEBUG			?= 0
 
 COMPONENT_SUBMODULES	:= axtls-8266
 
 COMPONENT_SRCDIRS := \
-  src \
+	src \
 	axtls-8266/compat \
 	axtls-8266/crypto \
 	axtls-8266/ssl
@@ -15,7 +15,7 @@ COMPONENT_SRCDIRS += \
 endif
 
 COMPONENT_INCDIRS := \
-  . \
+	. \
 	src \
 	axtls-8266 \
 	axtls-8266/ssl \
