+++ Sming/Components/ssl/NoSsl/FactoryImpl.cpp
@@ -1,0 +1,6 @@
+#include <Network/Ssl/Factory.h>
+
+namespace Ssl
+{
+Factory* factory = nullptr;
+}
