+++ Sming/Components/ssl/README.rst
@@ -1,0 +1,14 @@
+SSL
+===
+
+Provides Secure Socket Layer (SSL) support for Sming with selectable implementation.
+
+.. envvar:: ENABLE_SSL
+
+   1: (default) Enable SSL using AXTLS implementation
+   0: Disable SSL
+
+   Other values specify a supported SSL implementation:
+
+   - AXTLS
+
