+++ samples/Basic_Ssl/app/application.cpp
@@ -1,6 +1,5 @@
 #include <SmingCore.h>
 #include "Data/HexString.h"
-#include <AxtlsSsl.h>
 
 // If you want, you can define WiFi settings globally in Eclipse Environment Variables
 #ifndef WIFI_SSID
@@ -8,7 +7,6 @@
 #define WIFI_PWD "PleaseEnterPass"
 #endif
 
-AxtlsFactory sslFactory;
 Timer procTimer;
 HttpClient downloadClient;
 
@@ -95,7 +93,4 @@ void init()
 
 	WifiEvents.onStationGotIP(gotIP);
 	WifiEvents.onStationDisconnect(connectFail);
-
-	//
-	TcpConnection::setSslFactory(sslFactory);
 }
